chore(repo): docs, CI, lint pass, contributor scaffolding
- Add architecture/PII-policy/configuration/schemas/examples docs and
  acceptance-criteria evidence mapping for OpenAgriNet#1.
- Add CI workflow (lint, test on py3.10-3.12, end-to-end smoke, sdist+wheel
  build) and release workflow with PyPI trusted-publishing scaffolding.
- Add pre-commit config (ruff fix + format), GitHub issue/PR templates,
  CHANGELOG, CONTRIBUTING.
- Run ruff fix + format across the repo; flatten redundant nested ifs,
  use contextlib.suppress, add explicit zip strict=False, and tag the
  two file-handle-stored-on-self sites with noqa: SIM115. All 75 tests
  green; tp run --config configs/example.yaml passes end-to-end.

+++ src/training_pipeline/tagging/stratify.py
@@ -14,9 +14,8 @@ from __future__ import annotations
 import hashlib
 import random
 from collections import defaultdict
-from collections.abc import Iterable, Sequence
+from collections.abc import Sequence
 from dataclasses import dataclass
-from typing import Any
 
 from training_pipeline.schemas.events import Trajectory
 
@@ -87,8 +86,8 @@ def stratified_split(
         local = list(indices)
         rng.shuffle(local)
         n = len(local)
-        n_train = int(round(n * fractions[0]))
-        n_val = int(round(n * fractions[1]))
+        n_train = round(n * fractions[0])
+        n_val = round(n * fractions[1])
         # Anything left goes to test so totals add up exactly.
         n_test = max(0, n - n_train - n_val)
         train.extend(local[:n_train])
